separated the code into different files. and also, worked on the errors
