+++ src/redux/store.jsx
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit';
+import calculatorReducer from './calculatorSlice';
+
+const store = configureStore({
+  reducer: {
+    calculator: calculatorReducer,
+  },
+});
+
+export default store;
